# cmd/server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user.CreateUserRequestBody:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  util.Res:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:80
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server Petstore server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
...
  /user/list:
    get:
      consumes:
      - application/json
      description: List User
      operationId: ListUser
      parameters:
      - description: UserName
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      summary: ListUser
      tags:
      - ListUser
  /user/login:
    get:
      consumes:
      - application/json
      description: User Login
      operationId: Login
      parameters:
      - description: UserName
        in: query
        name: username
        required: true
        type: string
      - description: PassWord
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      summary: Login
      tags:
      - Login
  /user/register:
    post:
      consumes:
      - application/json
      description: Register User
      operationId: RegisterUser
      parameters:
      - description: Register User
        in: body
        name: ReqeustBody
        required: true
        schema:
          $ref: '#/definitions/user.CreateUserRequestBody'
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      summary: RegisterUser
      tags:
      - RegisterUser
  /user/verifytoken:
    get:
      consumes:
      - application/json
      description: Verify Token
      operationId: VerifyToken
      parameters:
      - description: token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: code":200,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "400":
          description: code":400,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "404":
          description: code":404,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
        "500":
          description: code":500,"data":null,"msg":""}
          schema:
            $ref: '#/definitions/util.Res'
      summary: VerifyToken
      tags:
      - VerifyToken
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"